+++ packages/plugin-gallery/web/src/consts.ts
@@ -1,12 +1,15 @@
 export const galleryModals = {
   settings: 'gallerySettings',
   manageMedia: 'manageMedia',
+  layout: 'galleryLayouts',
 };
 
 export const GALLERY_BUTTONS = {
   size: 'gallerySize',
   alignment: 'galleryAlignment',
   layout: 'galleryLayout',
+  mangeMedia: 'galleryMangeMedia',
+  addMedia: 'galleryAddMedia',
 };
 
 export const fileInputAccept = '.jpg,.png,.gif,.jpeg,.jpe,.jfif,.bmp,.heic,.heif,.tfif,.tif,.webp';