+++ packages/plugin-social-polls/web/src/toolbar/PollLayoutPanel.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { RicosContext } from 'ricos-context';
 import { KEYS_CHARCODE } from 'wix-rich-content-editor-common';
-import { ListItemSelect, DropdownModal } from 'wix-rich-content-toolbars-ui';
+import { ListItemSelect, DropdownModal, ListItemSection } from 'wix-rich-content-toolbars-ui';
 import { layoutData } from './PollButtonsData';
 
 type Props = {
@@ -17,31 +17,41 @@ const LineStylePanel: React.FC<Props> = ({
 }) => {
   const { t } = useContext(RicosContext) || {};
 
-  const onKeyDown = (e, commandKey) => {
+  const onKeyDown = (e, onClick) => {
     if (e.keyCode === KEYS_CHARCODE.ENTER) {
-      onClick(commandKey);
+      onClick();
       e.stopPropagation();
     }
   };
 
-  const DropdownOptions = layoutData.map(
-    ({ dataHook, icon: Icon, commandKey, text, component: Component }) =>
-      Component ? (
-        <Component text={t(text)} onClick={onCustomizeButtonClick} />
-      ) : (
-        <ListItemSelect
-          key={commandKey}
-          title={t(text)}
-          dataHook={dataHook}
-          prefix={<Icon />}
-          selected={commandKey === getSelectedLayout()}
-          onClick={() => onClick(commandKey)}
-          onKeyDown={e => {
-            onKeyDown(e, commandKey);
-          }}
-        />
-      )
+  const CustomizeButton = (
+    <>
+      <ListItemSection type={'divider'} />
+      <ListItemSelect
+        title={t('Poll_Mobile_Editor_Toolbar_Customize')}
+        onClick={onCustomizeButtonClick}
+        dataHook={'poll_layout_modal_button'}
+        onKeyDown={e => onKeyDown(e, onCustomizeButtonClick)}
+      />
+    </>
   );
+
+  const LayoutButtons = layoutData.map(({ dataHook, icon: Icon, text, commandKey }) => {
+    const onButtonClick = () => onClick(commandKey);
+    return (
+      <ListItemSelect
+        key={commandKey}
+        title={t(text)}
+        dataHook={dataHook}
+        prefix={<Icon />}
+        selected={commandKey === getSelectedLayout()}
+        onClick={onButtonClick}
+        onKeyDown={e => onKeyDown(e, onButtonClick)}
+      />
+    );
+  });
+
+  const DropdownOptions = [...LayoutButtons, CustomizeButton];
 
   return <DropdownModal options={DropdownOptions} />;
 };